Edited testing suite to be integrated with real CSV data
